+++ src/lib.rs
@@ -4,88 +4,10 @@ use nalgebra::{Const, DVectorView, DVectorViewMut, Dyn, SMatrix};
 use numpy::{PyArray2, PyReadonlyArray1, PyReadwriteArray1};
 use pyo3::prelude::*;
 use pyo3::types::PyDict;
+use crate::smallstrain::SmallStrainModel;
+use crate::smallstrain::linear_elastic::LinearElastic3D;
+pub mod smallstrain;
 
-struct LinearElastic3D {
-    D: SMatrix<f64, 6, 6>,
-}
-
-trait SmallStrainModel {
-    //fn new(parameters: &HashMap<String, f64>) -> Self;
-    fn evaluate_ip(
-        &self,
-        ip: usize,
-        del_t: f64,
-        stress: &mut DVectorViewMut<f64>,
-        del_strain: &DVectorView<f64>,
-        tangent: &mut DVectorViewMut<f64>,
-    );
-    fn evaluate(
-        &self,
-        del_t: f64,
-        stress: &mut DVectorViewMut<f64>,
-        del_strain: &DVectorView<f64>,
-        tangent: &mut DVectorViewMut<f64>,
-    ) {
-        assert_eq!(stress.nrows(), del_strain.nrows());
-        let n: usize = stress.nrows() / 6;
-        assert_eq!(stress.nrows(), n * 6);
-        for ip in 0..n {
-            self.evaluate_ip(ip, del_t, stress, del_strain, tangent);
-        }
-    }
-    fn evaluate_some(
-        &self,
-        del_t: f64,
-        stress: &mut DVectorViewMut<f64>,
-        del_strain: &DVectorView<f64>,
-        tangent: &mut DVectorViewMut<f64>,
-        ips: &[usize],
-    ) {
-        assert_eq!(stress.nrows(), del_strain.nrows());
-        let n: usize = stress.nrows() / 6;
-        assert_eq!(stress.nrows(), n * 6);
-        for ip in ips {
-            self.evaluate_ip(*ip, del_t, stress, del_strain, tangent);
-        }
-    }
-}
-
-/// implement the small strain model for linear elastic material
-impl SmallStrainModel for LinearElastic3D {
-
-    fn evaluate_ip(
-        &self,
-        ip: usize,
-        del_t: f64,
-        stress: &mut DVectorViewMut<f64>,
-        del_strain: &DVectorView<f64>,
-        tangent: &mut DVectorViewMut<f64>,
-    ) {
-        let mut view_stress = stress.fixed_view_mut::<6, 1>(ip * 6, 0);
-        let mut view_tangent = tangent.fixed_view_mut::<36, 1>(ip * 36, 0);
-        let view_strain = del_strain.fixed_view::<6, 1>(ip * 6, 0);
-        view_stress += self.D * view_strain;
-        view_tangent.copy_from_slice(&self.D.as_slice());
-    }
-}
-impl LinearElastic3D {
-    fn new(parameters: &HashMap<String, f64>) -> Self{
-        //let mut D = SMatrix::<f64, 6, 6>::zeros();
-        let E = parameters.get("E").unwrap();
-        let nu = parameters.get("nu").unwrap();
-        let mu = E / (2.0 * (1.0 + nu));
-        let lambda = E * nu / ((1.0 + nu) * (1.0 - 2.0 * nu));
-        let D = SMatrix::<f64, 6, 6>::new(
-            lambda + 2.0 * mu, lambda, lambda, 0.0, 0.0, 0.0,
-            lambda, lambda + 2.0 * mu, lambda, 0.0, 0.0, 0.0,
-            lambda, lambda, lambda + 2.0 * mu, 0.0, 0.0, 0.0,
-            0.0, 0.0, 0.0, 2.0 * mu, 0.0, 0.0,
-            0.0, 0.0, 0.0, 0.0, 2.0 * mu, 0.0,
-            0.0,0.0,0.0,0.0,0.0, 2.0 * mu,
-        );
-        Self { D: D }
-    }
-}
 
 /// Wrapper struct for LinearElastic3D in python
 #[pyclass(unsendable)]
@@ -231,16 +153,10 @@ impl PyLinearElastic3D {
         Ok(())
     }
 }
-/// Formats the sum of two numbers as string.
-#[pyfunction]
-fn sum_as_string(a: usize, b: usize) -> PyResult<String> {
-    Ok((a + b).to_string())
-}
-
+/// 
 /// A Python module implemented in Rust.
 #[pymodule]
 fn comfe(_py: Python, m: &PyModule) -> PyResult<()> {
-    m.add_function(wrap_pyfunction!(sum_as_string, m)?)?;
     m.add_class::<PyLinearElastic3D>()?;
     m.add_class::<PySmallStrainModel>()?;
     m.add_function(wrap_pyfunction!(py_new_linear_elastic_3d, m)?)?;
